Strong neutral color.
