+++ index.html
@@ -73,6 +73,20 @@
               {% include swatch.html id="accentNonContentStrong" %}
             </figure>
           </div>
+          <div class="row ui-swatches">
+            <figure class="color-swatch">
+              <figcaption>Neutral Non-Content Faint</figcaption>
+              {% include swatch.html id="neutralNonContentFaint" %}
+            </figure>
+            <figure class="color-swatch">
+              <figcaption>Neutral Non-Content Subdued</figcaption>
+              {% include swatch.html id="neutralNonContentSubdued" %}
+            </figure>
+            <figure class="color-swatch">
+              <figcaption>Neutral Non-Content Strong</figcaption>
+              {% include swatch.html id="neutralNonContentStrong" %}
+            </figure>
+          </div>
         </div>
 
         <div id="contentGroup" class="col">
@@ -88,6 +102,16 @@
             <figure class="color-swatch">
               <figcaption>Accent Content Strong</figcaption>
               {% include swatch.html id="accentContentStrong" %}
+            </figure>
+          </div>
+          <div class="row ui-swatches">
+            <figure class="color-swatch">
+              <figcaption>Neutral Content Subdued</figcaption>
+              {% include swatch.html id="neutralContentSubdued" %}
+            </figure>
+            <figure class="color-swatch">
+              <figcaption>Neutral Content Strong</figcaption>
+              {% include swatch.html id="neutralContentStrong" %}
             </figure>
           </div>
         </div>
